+++ sso/settings_test.go
@@ -61,14 +61,6 @@ func (suite *SettingsTestSuite) TestLoadSettings() {
 	t := suite.T()
 
 	assert.Equal(t, TEST_SETTINGS_FILE, suite.settings.ConfigFile())
-
-	// ensure we upgraded ConfigUrlAction to ConfigProfilesUrlAction
-	assert.Equal(t, "", suite.settings.ConfigUrlAction)
-	assert.Equal(t, url.ConfigProfilesOpen, suite.settings.ConfigProfilesUrlAction)
-
-	// ensure we upgraded FirefoxOpenUrlInContainer
-	assert.False(t, suite.settings.FirefoxOpenUrlInContainer)
-	assert.Equal(t, url.OpenUrlContainer, suite.settings.UrlAction)
 }
 
 func (suite *SettingsTestSuite) TestGetSelectedSSO() {
@@ -331,30 +323,7 @@ func TestCreatedAt(t *testing.T) {
 }
 
 func TestApplyDeprecations(t *testing.T) {
-	s := &Settings{
-		ListFields:                []string{"Foo", "Bar", "ExpiresStr", "AccountIdStr", "ARN"},
-		ProfileFormat:             "{{ AccountIdStr .AccountId }}:{{ .RoleName }}",
-		FirefoxOpenUrlInContainer: true,
-		ConfigProfilesUrlAction:   url.ConfigProfilesUndef,
-		ConfigUrlAction:           string(url.Exec),
-	}
-
-	r := s.applyDeprecations()
-	assert.True(t, r)
-
-	// Upgrade ConfigUrlAction to ConfigProfilesUrlAction
-	assert.Equal(t, string(url.Undef), s.ConfigUrlAction)
-	assert.Equal(t, url.ConfigProfilesExec, s.ConfigProfilesUrlAction)
-
-	// Upgrade FirefoxOpenUrlInContainer to UrlAction
-	assert.Equal(t, url.OpenUrlContainer, s.UrlAction)
-	assert.Equal(t, false, s.FirefoxOpenUrlInContainer)
-
-	// ExpiresStr => Expires, etc
-	assert.Equal(t, []string{"Foo", "Bar", "Expires", "AccountIdPad", "Arn"}, s.ListFields)
-
-	// AccountIdStr .AccountId => .AccountIdPad
-	assert.Equal(t, "{{ .AccountIdPad }}:{{ .RoleName }}", s.ProfileFormat)
+	// as of v2.0 it does nothing...
 }
 
 func TestGetExecutable(t *testing.T) {